Refactor thermal interface and material properties
- Updated `thermal_interface.F90` to streamline initialization and computation procedures for thermal properties.
- Introduced new elemental subroutines for calculating solid density and specific heat in `materials.F90`.
- Modified `heat_capacity_multiphase.F90` to adjust vapor heat capacity calculation.
- Enhanced `materials_base.F90` with new methods for retrieving solid properties.
- Added latent heat calculation methods in `models.F90` and `phase_systems.F90`.
- Implemented pressure derivative calculations for ice and water phases in `gcc_interface.F90` and `gcc_base.F90`.
- Integrated IAPWS wrapper for water and ice models in `physics_service.F90` and `physics_types.F90`.
- Updated test program to include new domain features and display finite element information.

+++ src/domain/fe/fe.F90
@@ -89,7 +89,7 @@ contains
         !> The FE manager object
         class(type_fe_manager), intent(in), target :: self
         !> The ID of the FE object
-        integer, intent(in) :: fe_id
+        integer(int32), intent(in) :: fe_id
         !> Pointer to the requested FE object
         class(abst_fe), pointer :: fe
